wpa_supplicant: Send P2P mgmt iface add/remove notifications
wpa_supplicant has 2 different |wpa_supplicant| structure entries in
|global->ifaces| for the virtual p2p interface (wpa_s->p2p_mgmt == 1)
and the underlying station mode interface.
However, the control IPC's (dbus, hidl) are not currently
notified about the addition of the virtual p2p interface because of the
|if (wpa_s->p2p_mgmt == 0)| check in |wpa_supplicant_add_iface|.
For the dbus control, we didn't want to track the existence of 2
separate ifaces (always used global->p2p_init_wpa_s| to reach the P2P
iface. But for HIDL control, we want to expose the P2P iface object
when the corresponding |wpa_supplicant| entry is added to
|global->ifaces|.
So, send notifications to indicate the addition/removal of
the p2p mgmt iface to HIDL.

Test: Compiles
Bug: 31497295
Change-Id: Iad63521565f58e2e1ec871046d0b0d65f271486e