wpa_supplicant(hidl): Fix network object removal
Was attempting to remove the network object from the wrong object map
(iface instead of network) in HidlManager on network removal.

While there,
Add some error logs for these object management errors.

PS: The network object not being removed in the wpa_supplicant's service
is not detectable since the next network created with the same network Id
would reuse the same object. But, this results in us not clearing the
callback list on the supplicant's service side and
will leave behind stale callbacks. This combined with the known
limitation about java HIDL callback objects not being refcounted (b/33042939),
will eventually result in a crash when one of the service invokes one of the
stale callbacks that was garbage collected in framework. This is the reason the
crash is only seen in EAP SIM tests because the network callbacks are only used
for such networks.

Bug: 35703650
Test: Compiles
Change-Id: I13c50245f059e330b1d7ef521d309c0c3b366f13