wpa_supplicant(hidl): remove unused parameter
The |ip| parameter for the P2P group started event is
unused. Despite that, we check if the |ip| is null.
And, if |ip| is null, we don't report the event up
to framework.

Note that
- we never actually report the |ip| to framework
- some listeners want to know about group start events, even
  when the |ip| is null.

Given the above, the null check is unnecessary. Since
the null check is unnecessary, the |ip| itself is
unnecessary.

Hence: remove the |ip| parameter for this event.

Bug: 37286961
Test: Compiles
Change-Id: I89c22d7513fd582e396057b771da33a13345f54b