Prevent reinstallation of an already in-use group key
Track the current GTK and IGTK that is in use and when receiving a
(possibly retransmitted) Group Message 1 or WNM-Sleep Mode Response, do
not install the given key if it is already in use. This prevents an
attacker from trying to trick the client into resetting or lowering the
sequence counter associated to the group key.

Change-Id: Iabccd7aa958126965f5d666dcc145622526e42cf
Signed-off-by: Mathy Vanhoef <Mathy.Vanhoef@cs.kuleuven.be>
Signed-off-by: Glen Kuhne <kuh@google.com>
Merged-In: I1273cf9ee58154b16512ef4284a8d2c5541d3c01
Merged-In: I54f22e16b4fca5c541fd7f7e6cf6c9503e875f4e
Bug: 65245581
Test: WiFi integration tests