Create a hostapd build target without the HIDL
Create a build target that does not incldue the HIDL control interface.
This is useful for emulators that use hostapd as a way to provide a
virtual access point while still allowing them to have a hostapd binary
with the HIDL control interface that can be used to support WiFi
hotspots.

Note: this cl does not impact real devices, as the target
device 'generic_x86, generic_x86_64, generic_arm and generic_arm64'
refer to emulator devices.
in addition, this will only impact emulator's vendor image

BUG: 74401469
Test: run cts -m CtsNetTestCases
Change-Id: I76a6ba18808bb2e2bb8a7f8978acdae71de6c325