Remove LOCAL_MODULE_TAGS := debug
This should be replaced with adding these modules to
PRODUCT_PACKAGES_DEBUG in the appropriate product makefiles instead. In
this case, hostapd and wpa_supplicant are added to PRODUCT_PACKAGES in
various device makefiles, many of them already specifying the _cli
variants at the same time.

This does not affect user builds.

Test: treehugger
Change-Id: Iba97d297b440c03b639ba5691e1d954f7d51a2eb