[wpa_supplicant] Fix a regression in storing of external_auth SSID/BSSID
An earlier change in drivers_ops API for struct external_auth broke the
way SSID and BSSID for an external authentication request were stored.
The implementation depended on the memory array being available in the
API struct with a use of memcpy() to copy the full structure even though
when only SSID and BSSID was needed. Fix this by replacing that
easy-to-break storing mechanism with explicit arrays for the exact set
of needed information.

Bug: 133798691
Test: Associate/disassociate to/from WPA3 SAE APs
Cherry-picked from: d42df8d6ce81d47aea8059c45e3db5c51897f7e8

Change-Id: I02e2703a3b0a8b6dd3163783a196ded4b8f200d3