Do not clear anonymous id when anonymous@<realm> is used
If the EAP-SIM/AKA server does not provide a new pseudonym and the
locally configured "pseudonym" in anonymous_identity is actually
an anonymous identity instead of a real EAP-SIM/AKA pseudonym, do
not clear the anonymous_identity network profile parameter.
This is needed to avoid forgetting the anonymous identity when
going through EAP-SIM/AKA authentication and then reverting back
to using IMSI-based (e.g., encrypted) identity.

Bug: 138610772
Test: Associate to <carrier> AP, get out of range, reassociate
Test: Associate to <carrier> AP, pop SIM, reinsert, reassociate
Change-Id: I511575aa0bf5e43809f0bcea756822f221453a83