Write the op_class in hostapd.conf file
With the addition of 6GHz band, channel number is no longer
sufficient to identify a channel since 6GHz band channels use
same numbers as 2.4/5GHz bands. Hence, for SoftAp with no ACS,
the channel number in hostapd.conf needs to be coupled with an
op_class attribute.

This commit calculates the op_class for the channel number, and
pass it in the hostapd.conf.

Bug: 146186687
Test: Manual
Change-Id: I69ee3cc89c19d825eb5076b0a14f770a1385a859