p2p: route events to P2P device
The P2P device interface, when available, is used
for all P2P management operations, and all P2P events
are sent to it. The supplicant already routes all
relevant events to the P2P device but some events
are triggered on the network interface(group interface)
and not routed properly to the P2P device.
Fix the HIDL code to send these events on the P2P
device interface if found.

Bug: 146760186
Test: atest VtsHalWifiSupplicantV1_0TargetTest \
            VtsHalWifiSupplicantP2pV1_0TargetTest \
            VtsHalWifiSupplicantV1_1TargetTest \
            VtsHalWifiSupplicantV1_2TargetTest \
            VtsHalWifiSupplicantP2pV1_2TargetTest \
            VtsHalWifiSupplicantV1_3TargetTest
Change-Id: I1f21e14ec67253192971a99effb8cf07d265e849