p2p: enable auto join to join a group if possible
During search state, the probe response from a GO
peer won't mark GROUP_OWNER bit. This might overrie
the correct device group capabilities. In the meantime,
there might be also timing issue which means that a
GO peer information is not updated in time when the third
client try to join the group.

If the peer does not know the target is a GO,
it would try starting group negotiation flow, but
won't initiate the WPS flow to join the group.

Enabling auto join could ensure that every peer would
try to scan the group first before starting the group
negotiation flow.

This will increase additional 4 - 5 seconds (a full scan time)
before starting the negotiation flow.

Bug: 163532173
Test: atest VtsHalWifiSupplicantV1_0TargetTest \
            VtsHalWifiSupplicantV1_1TargetTest \
            VtsHalWifiSupplicantV1_2TargetTest \
            VtsHalWifiSupplicantV1_3TargetTest \
            VtsHalWifiSupplicantV1_4TargetTest
Test: act.py
          -c tools/test/connectivity/acts/tests/google/wifi/p2p/config/wifi_p2p.json \
          -tb WifiP2pAllAttached \
          -tc WifiP2pMultiPeersTest WifiP2pLocalServiceTest WifiP2pManagerTest
Test: act.py
          -c tools/test/connectivity/acts/tests/google/wifi/p2p/config/wifi_p2p_group.json \
          -tb WifiP2pAllAttached \
          -tc WifiP2pGroupTest

Change-Id: I6cc3c704d70939d2fb61824467de31b1c480c62b