hostapd: Fix eapol packet drop in bridged mode
When AP interface needs to do scan for some reason, for instance: HT
SCAN for bandwidth check. It needs to configure interface mode to STA.
In bridged mode, the STA interface doesn't allow to link to bridged
interface. In current hostapd design, it will remove bridged link in
temporary. And add it back when scan done. But it doesn't add ifindex
link back between bridged interface and AP interface. It will cause
eapol rx packet drop in hostapd because eapol packet is received from
bridged interface. And eapol rx handling will check if ifindex exist
in current hostapd but it doesn't exist because it was removed when
re-configure mode to STA.

Fix: Add ifindex link back when do reset to AP mode.

Bug: 183577766
Test: Manual Test, client can connect to wpa2 bridged AP.
Change-Id: Ide9e110f1b208e175ccc58f1f23c4b2084e38fa3