optimize the p2p client scan time
1. Reduced the total number of scans to find an auto GO
from 10 to 3.
2. Optimized the overall scan duration for band scan, ie,
   a. In the first two scans, include all the channels except DFS channels.
   c. In the third attempt, include all the channels.
3. When framework gives a specific frequency to connect,
no need to check if the frequency is allowed for p2p operation.
Instead check the frequency is allowed as per the regulatory. If the channel is not
allowed to operate, return failure to application. This is to allow devices to scan passive
channel and connect(GO takes the master responsibilities).

Bug: 186675298
Test: File transfer using nearby APP
Change-Id: Icf3494d4c655d85ce9f67e4f90632022303a7e75