Use a shorter wait time for p2p connect with operating frequency
If supplicant is receiving the connect request with GO operating frequency,
then application(caller) must be already knowing and sure that GO is
operating on that frequency. So if the GO is not seen in the first
scan attempt, it is not worth waiting for 1 second to trigger the next scan.
It's better to finish scans quickly and report the Join failure to application.
This will help applications to trigger another Join attempt with a band information
or without band or frequency information.
To speed up the scan, reduced the scan interval for single channel scan
from 1 sec to 200 milli seconds.

Bug: 191718156
Test: Manual - Triggered P2P join request with channel, band, no
channel/band and verified the scan times.

Change-Id: I44e22f8e3a4d7460466b3e102f2d0c92038914d1