set ieee80211w to optional in hostapd config
set ieee80211w=1 config item in the hostapd config file.
This is to enable IEEE 802.11 management frame protection
feature optional on SoftAp interface. This is mainly to
prevent denial of service when an attacker sends deauthentication
frames to the clients connected to the SoftAp.

Bug: 197874458
Test: Connected STAs to Pixel SofAp and confirmed from the
sniffer logs that PMF capabilities are exchanged and the required
management frames are protected.

Change-Id: Ia773b79a65771a7b53c561ed2601f02aee82f0ea