Don't set he_oper_chwidth for 2.4GHz only band
Channel width of 80MHz is not supported in the 2.4GHz band.
So do not set he_oper_chwidth configuration parameter
in the hostapd configuration file for 2.4GHz only mode.
This is to avoid failure in selecting the channel in
hostapd based ACS algorithm.

Bug: 240579469
Test: Manual - Able to bring up SoftAp in 2.4GHz and verified
      the hostapd config file.
Change-Id: I76a5283a4f4397f4b65420b8a9cd0455249c50f1
Merged-In: I76a5283a4f4397f4b65420b8a9cd0455249c50f1