P2P: support preferred GO band
Allow specifying preferred GO band in addition to frequency. If a band
is specified, the first two scans will be limited to only DFS channels
to shorten scan times, and the next two will scan the entire band.

Bug: 233925359
Test: Used WifiP2PTest app to specify band and looked at logs to confirm
that first two scans were non-radar and third was radar included. Also
verified that specificying a single channel still works - all scans are
on that channel.

Change-Id: I2d558bb2c00bf3925ad92f5f036335cd46b9c609