wifi: Add the connected frequency in the state change event
Currently framework gets the connected AP frequency
information by checking the scan cache & through signal
poll command. But these mechanisms doesn't work if framework
scan cache is not updated with the connected AP information
(For example in case of roaming and force connection without scan).
To fix this modified supplicant state change interface
and included the frequency of the connected channel in the
state change event.

Bug: 240299862
Test: Manual - Tested STA - AP connection and verfied the data
      from debug logs.
Change-Id: I5b6b375edea1bc645d0745bb0d718322917ad585