Allow compilation without HS20/INTERWORKING
CONFIG_HS20 and CONFIG_INTERWORKING are used in wpa_supplicant
to enable or disable HotSpot 2.0 or ANQP respectively. These
defines were not checked in AIDL code, which assumed that
HS20 and INTERWORKING were enabled, and fails to compile when
they are disabled.

This patch compiles out the functions that require HS20 and
INTERWORKING when they are not enabled in android.config.

Upstream from Meta.

Test: Modify android.config and build for all configurations of
CONFIG_HS20 and CONFIG_INTERWORKING. WiFi continues to work fine.

Bug: 280661640

Signed-off-by: Sam Gao <ssgao@meta.com>

Change-Id: Ib95b798ea80c375f6922b4b46576578ad7c112c2