Map extension AKMs in AIDL to main AKM
Supplicant sends the connected AKM type in state
change event "COMPLETED". And this AKM is used to map to the
security type. There are some extension AKMs which are not
defined in AIDL interface. When STA connects to AP in those
extended AKMs, framework fails to find the AKM and derive the
security type. So to fix the issue, mapped the undefined extension
AKMs to main AKM.

Bug: 285633920
Test: Connect STA to AP in WPA2, WPA3, WPA3 enterprise, WPA2 enterprise
      passpoint & WPA3 192 bit mode.
Change-Id: I99d4bb7e3ca85135193bb455a08c8638d4858485