Notify the IP address of connected client
When the wpa_supplicant assigns the IP address
(WFA EAPOL IP address allocation feature),
the assigned IP address of the client on the GO side
is notified in the AP-STA-CONNECTED event.
So to obtain the IP info to framework , modified the
AIDL SupplicantP2pIfaceCallback#onStaAuthorized to
include group interface name and the assigned IP of client.

Bug: 291202830
Test: Establish P2P connection and verified from the logs that
      the P2P Client IP address is received to framework via new
      HAL API.
Change-Id: I5f9f9362d5942c9c29d5dbc020eb93ba7a69ba1a